+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 4.0.0 (tbd)
 
    - New auto-build system for franka.mtlbx under github release page. 
-   - Default build --> libfranka 0.18.1.
+   - Default build --> libfranka 0.18.2.
 
 ## 3.1.0 (20-10-2025)